Change setState syntax
While a small thing (and not a breaking change for the lab in question) onChangeType uses setState in a way that could have setState updating this.state.filters with old data. It's my understanding, from previous lessons, that it is preferable to pass state via an arrow function rather than use this.state directly. Please correct me if I'm wrong, as well this is my first change proposal so don't bite :)